SmartHelpdesk/requirement.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
scikit-learn==1.5.2
python-dotenv==1.0.1
requests==2.32.3